Rework to get this demo working again
* Remove Soundpipe dependency, as Paul is no longer maintaining
  this library; this also eliminates the need for libsndfile

* Add minimal set of Soundpipe definitions to allow Voc to compile
  without modification

* Use local copy of Voc, in lieu of the one embedded in Soundpipe

* On Linux, configure RtAudio to use ALSA instead of JACK
  (drawing on a commit in Orhan Kavrakoğlu's fork)

* Link with $(CXX) instead of "$(CC) -lstdc++"

* Fix minor compiler warnings

NOTE: The README.md file needs updating!

+++ main.c
@@ -163,7 +163,7 @@ int main(int argc, char *argv[])
             NK_WINDOW_MINIMIZABLE|NK_WINDOW_TITLE))
         {
             nk_layout_row_static(ctx, 30, 200, 1);
-            for(i = 0; i < vd.tract_size; i++) {
+            for(i = 0; i < (int)vd.tract_size; i++) {
                 nk_slider_float(ctx, 0, &vd.tract[i], 3.5, 0.01);
             }
         }
